Finished Video 3, contactClass was made and methods for Select, Add, Delete, and Update were created.
I followed what the video showed, though there's a large amount of repeated code that can be cut out into a method to reduce that. I also don't like the raw SQL, so I may edit that once the program is all running to follow the best conventions I've learned in school.