Program runs, picks winners, and opens html file.
The html file doesn't actually have anything useful to it yet, so that's the next step. It was a pain to get the html file even to be discovered though. The method which I found to do this, wasn't working because it seems to think that the path for the program is in an entirely different location. I set it up to remove the portion of the directory that isn't correct, so hopefully this will work no matter what when I get it on another computer. We shall see.