Creating a more appealing look for the print out.
After the main bulk of the program was created, it was decided that a more appealing display was desired. Now the names are displaying in a centered div with some style to make it look nicer. More adjustments will be happening, but this is the point that I finally got the table to display how I wanted, and moved to a table from just divs because it wasn't working how I wanted.