Added the draw date to the html page.
In addition to this, I modified a little styling to make some parts look a little nicer and spread out. Took a lot to get something as simple as padding to work because of the way I had to set up the whole table to display, but it looks nice now and is what was requested for the display.