Changed the message box to actually be a timer on its own to avoid a bug with the HTML page not being reset.
This is kind of a weird change that was necessary because of the use of an HTML page that is getting edited in place and then reset. I wanted to remove the user clicked message box, but with the speed of the program, it was possible to skip resetting the HTML code back to what it should have been. I tried a bunch of ways of doing this, and the final one I found was to just make another temp form that auto closes.