Edited the child window to behave as such.
The add window now pops up in the middle of the main program window no matter where it is. This is a nice feature, and something I have struggled with before doing other things where I had a pop up and it always went to a weird spot. Wish I had this type of thing before.